backend/requirements.txt
altgraph==0.17.4
blinker==1.9.0
bottle==0.13.2
cffi==1.17.1
click==8.1.7
clr-loader==0.2.6
colorama==0.4.6
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
packaging==24.2
pefile==2023.2.7
proxy_tools==0.1.0
pycparser==2.22
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
pythonnet==3.0.4
pywebview==5.3.2
pywin32-ctypes==0.2.3
setuptools==75.6.0
typing_extensions==4.12.2
Werkzeug==3.1.3
